+++ main-data/src/main/java/com/teamhy2/main/data/di/MainModule.kt
@@ -3,9 +3,9 @@ package com.teamhy2.main.data.di
 import com.teamhy2.main.data.repository.DefaultStudyDayRepository
 import com.teamhy2.main.data.repository.DefaultWebViewRepository
 import com.teamhy2.main.data.repository.DefaultWiseSayingRepository
-import com.teamhy2.main.domain.StudyDayRepository
-import com.teamhy2.main.domain.WebViewRepository
-import com.teamhy2.main.domain.WiseSayingRepository
+import com.teamhy2.main.domain.repository.StudyDayRepository
+import com.teamhy2.main.domain.repository.WebViewRepository
+import com.teamhy2.main.domain.repository.WiseSayingRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
