--- main-domain/src/main/java/com/teamhy2/main/domain/WiseSayingRepository.kt
@@ -1,7 +1,0 @@
-package com.teamhy2.main.domain
-
-import com.teamhy2.main.model.WiseSaying
-
-interface WiseSayingRepository {
-    suspend fun fetchWiseSaying(): WiseSaying
-}
