+++ main-domain/src/main/java/com/teamhy2/main/domain/repository/WebViewRepository.kt
@@ -1,4 +1,4 @@
-package com.teamhy2.main.domain
+package com.teamhy2.main.domain.repository
 
 interface WebViewRepository {
     suspend fun fetchFirebaseUrls(): Map<String, String>
