+++ main-domain/src/main/java/com/teamhy2/main/domain/repository/WiseSayingRepository.kt
@@ -1,0 +1,7 @@
+package com.teamhy2.main.domain.repository
+
+import com.teamhy2.main.domain.model.WiseSaying
+
+interface WiseSayingRepository {
+    suspend fun fetchWiseSaying(): WiseSaying
+}
